Closes #7 - configurable authorization enabled/disabled
