+++ alignak_module_ws/etc/arbiter/modules/mod-ws.cfg
@@ -52,6 +52,9 @@ define module {
     # Do not comment the port parameter (see Alignak #504)
     port                    8888
 
+    # HTTP authorization
+    authorization           1
+
     # SSL configuration
     use_ssl                 0
     #ca_cert                 /usr/local/etc/alignak/certs/ca.pem
