+++ test/test_module.py
@@ -421,7 +421,7 @@ class TestModuleWs(AlignakTest):
             re.escape("configuration, listening on: http://me:8080"), 7)
 
     def test_module_zzz_basic_ws(self):
-        """Test the module basic API
+        """Test the module basic API - authorization enabled
 
         :return:
         """
@@ -568,6 +568,82 @@ class TestModuleWs(AlignakTest):
         for endpoint in api_list:
             print("Trying %s" % (endpoint))
             response = requests.get('http://127.0.0.1:8888/' + endpoint, auth=auth)
+            print("Response %d: %s" % (response.status_code, response.content))
+            self.assertEqual(response.status_code, 200)
+            if response.status_code == 200:
+                print("Got %s: %s" % (endpoint, response.json()))
+            else:
+                print("Error %s: %s" % (response.status_code, response.content))
+
+        self.modulemanager.stop_all()
+
+    def test_module_zzz_unauthorized(self):
+        """Test the module basic API - authorization disabled
+
+        :return:
+        """
+        self.print_header()
+        # Obliged to call to get a self.logger...
+        self.setup_with_file('cfg/cfg_default.cfg')
+        self.assertTrue(self.conf_is_correct)
+
+        # -----
+        # Provide parameters - logger configuration file (exists)
+        # -----
+        # Clear logs
+        self.clear_logs()
+
+        # Create an Alignak module
+        mod = Module({
+            'module_alias': 'web-services',
+            'module_types': 'web-services',
+            'python_name': 'alignak_module_ws',
+            # Alignak backend
+            'alignak_backend': 'http://127.0.0.1:5000',
+            'username': 'admin',
+            'password': 'admin',
+            # Set Arbiter address as empty to not poll the Arbiter else the test will fail!
+            'alignak_host': '',
+            'alignak_port': 7770,
+            'authorization': '0'
+        })
+
+        # Create the modules manager for a daemon type
+        self.modulemanager = ModulesManager('receiver', None)
+
+        # Load an initialize the modules:
+        #  - load python module
+        #  - get module properties and instances
+        self.modulemanager.load_and_init([mod])
+
+        my_module = self.modulemanager.instances[0]
+
+        # Clear logs
+        self.clear_logs()
+
+        # Start external modules
+        self.modulemanager.start_external_instances()
+
+        # Starting external module logs
+        self.assert_log_match("Trying to initialize module: web-services", 0)
+        self.assert_log_match("Starting external module web-services", 1)
+        self.assert_log_match("Starting external process for module web-services", 2)
+        self.assert_log_match("web-services is now started", 3)
+
+        # Check alive
+        self.assertIsNotNone(my_module.process)
+        self.assertTrue(my_module.process.is_alive())
+
+        time.sleep(1)
+
+        # Get the module API list and request on each endpoint
+        response = requests.get('http://127.0.0.1:8888')
+        print("Response: %s" % response)
+        assert response.status_code == 200
+        api_list = response.json()
+        for endpoint in api_list:
+            print("Trying %s" % (endpoint))
+            response = requests.get('http://127.0.0.1:8888/' + endpoint)
             print("Response %d: %s" % (response.status_code, response.content))
             self.assertEqual(response.status_code, 200)
             if response.status_code == 200:
